Move IntersectByFrequency into pipeline_utils and setup v2
